Adds CompletePurchase google directive (#234)
* Adds CompletePurchase google directive

* Adds Digital Goods docs

* Removes git-legal from codeclimate

* Makes transactionOptions parameter optional

* Adds PURCHASE_STATUS enum

* Changes to v3.2.0
