+++ .codeclimate.yml
@@ -7,8 +7,6 @@ exclude_patterns: # Exclude files and/or directories from analysis
 plugins:
   nodesecurity:
     enabled: true
-  git-legal:
-    enabled: true
   shellcheck:
     enabled: true
   fixme:
