+++ src/platforms/dialogflow/google/apis/index.ts
@@ -1,0 +1,2 @@
+export { DigitalGoods } from "./DigitalGoods";
+export { ITransactionOptions } from "./ITransactionOptions";
